move hardhat networks config to helpers
